requirements.txt
pyodbc==5.2.0
python-dotenv==1.0.1